Move XOTransport, pull node from XO in endpoint layer, if used
