+++ pyon/net/endpoint.py
@@ -21,7 +21,7 @@ from pyon.net.channel import ChannelClosedError, PublisherChannel, ListenChannel
 from pyon.core.interceptor.interceptor import Invocation, process_interceptors
 from pyon.util.containers import get_ion_ts, get_ion_ts_millis
 from pyon.util.log import log
-from pyon.net.transport import NameTrio, BaseTransport
+from pyon.net.transport import NameTrio, BaseTransport, XOTransport
 from pyon.util.sflow import SFlowManager
 
 # create special logging category for RPC message tracking
@@ -354,7 +354,20 @@ class SendingBaseEndpoint(BaseEndpoint):
         # ensure NameTrio
         self._send_name = self._ensure_name_trio(self._send_name)
 
+        # set node from XO
+        if isinstance(self._send_name, XOTransport):
+            if self.node is not None and self.node != self._send_name.node:
+                log.warn("SendingBaseEndpoint.__init__: setting new node from XO")
+
+            self.node = self._send_name.node
+
     def create_endpoint(self, to_name=None, existing_channel=None, **kwargs):
+        if to_name is not None and isinstance(to_name, XOTransport):
+            if self.node is not None and self.node != to_name.node:
+                log.warn("SendingBaseEndpoint.create_endpoint: setting new node from XO when node already set")
+
+            self.node = to_name.node
+
         e = BaseEndpoint.create_endpoint(self, to_name=to_name, existing_channel=existing_channel, **kwargs)
 
         name = to_name or self._send_name
@@ -469,6 +482,13 @@ class ListeningBaseEndpoint(BaseEndpoint):
 
         # ensure NameTrio
         self._recv_name = self._ensure_name_trio(self._recv_name)
+
+        # set node from XO
+        if isinstance(self._recv_name, XOTransport):
+            if self.node is not None and self.node != self._recv_name.node:
+                log.warn("ListeningBaseEndpoint.__init__: setting new node from XO when node already set")
+
+            self.node = self._recv_name.node
 
         self._ready_event = event.Event()
         self._binding = binding
