+++ pyon/net/transport.py
@@ -1028,3 +1028,18 @@ class LocalTransport(BaseTransport):
     def purge_impl(self, queue):
         return self._broker.purge(queue)
 
+class XOTransport(ComposableTransport):
+    """
+    Base Transport object for Exchange Objects.
+    """
+    def __init__(self, exchange_manager, priviledged_transport, node):
+        self._exchange_manager = exchange_manager
+        self.node              = node
+
+        ComposableTransport.__init__(self, priviledged_transport, None, *ComposableTransport.common_methods)
+
+    def setup_listener(self, binding, default_cb):
+        log.debug("XOTransport passing on setup_listener")
+        pass
+
+
